+++ R/RcppExports.R
@@ -17,12 +17,24 @@ rcpp_null_to_na <- function(x) {
     invisible(.Call(`_jsonify_rcpp_null_to_na`, x))
 }
 
-rcpp_df_to_json <- function(df, unbox = FALSE) {
-    .Call(`_jsonify_rcpp_df_to_json`, df, unbox)
+rcpp_pretty_json <- function(json) {
+    .Call(`_jsonify_rcpp_pretty_json`, json)
 }
 
-rcpp_numeric_to_json <- function(nv, unbox = FALSE) {
-    .Call(`_jsonify_rcpp_numeric_to_json`, nv, unbox)
+rcpp_minify_json <- function(json) {
+    .Call(`_jsonify_rcpp_minify_json`, json)
+}
+
+rcpp_pretty_print <- function(json) {
+    invisible(.Call(`_jsonify_rcpp_pretty_print`, json))
+}
+
+rcpp_df_to_json <- function(df, unbox = FALSE, digits = -1L) {
+    .Call(`_jsonify_rcpp_df_to_json`, df, unbox, digits)
+}
+
+rcpp_numeric_to_json <- function(nv, unbox = FALSE, digits = -1L) {
+    .Call(`_jsonify_rcpp_numeric_to_json`, nv, unbox, digits)
 }
 
 rcpp_character_to_json <- function(sv, unbox = FALSE) {
@@ -37,23 +49,23 @@ rcpp_logical_to_json <- function(iv, unbox = FALSE) {
     .Call(`_jsonify_rcpp_logical_to_json`, iv, unbox)
 }
 
-rcpp_list_to_json <- function(lst, unbox = FALSE) {
-    .Call(`_jsonify_rcpp_list_to_json`, lst, unbox)
+rcpp_list_to_json <- function(lst, unbox = FALSE, digits = -1L) {
+    .Call(`_jsonify_rcpp_list_to_json`, lst, unbox, digits)
 }
 
 rcpp_integer_matrix_to_json <- function(m, unbox = FALSE) {
     .Call(`_jsonify_rcpp_integer_matrix_to_json`, m, unbox)
 }
 
-rcpp_numeric_matrix_to_json <- function(m, unbox = FALSE) {
-    .Call(`_jsonify_rcpp_numeric_matrix_to_json`, m, unbox)
+rcpp_numeric_matrix_to_json <- function(m, unbox = FALSE, digits = -1L) {
+    .Call(`_jsonify_rcpp_numeric_matrix_to_json`, m, unbox, digits)
 }
 
 rcpp_character_matrix_to_json <- function(m, unbox = FALSE) {
     .Call(`_jsonify_rcpp_character_matrix_to_json`, m, unbox)
 }
 
-rcpp_validate_json <- function(geojson) {
-    .Call(`_jsonify_rcpp_validate_json`, geojson)
+rcpp_validate_json <- function(json) {
+    .Call(`_jsonify_rcpp_validate_json`, json)
 }