+++ R/jsondf-package.R
@@ -1,3 +1,7 @@
 #' @useDynLib jsonify, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
 NULL
+
+
+#' @export
+print.json <- function( x, ... ) cat( x )