+++ NEWS.md
@@ -1,4 +1,11 @@
-# jsonify 0.1.2
+## v0.1.3
 
-* `auto_unbox` argument for `to_json()`
+* `print.json()` method to print json class objects
+* `minify_json()` functino to remove indentation
+* `pretty_json()` function to add indentation
+* `digits` argument to specify the number of digits to which numeric values will be rounded
+
+## v0.1.2
+
+* `unbox` argument for `to_json()`
 * Added a `NEWS.md` file to track changes to the package.